Add data credit to navigator
