+++ react/src/navigation/navigator.tsx
@@ -2,6 +2,7 @@ import React, { createContext, ReactNode, useContext, useEffect, useMemo, useSta
 import { z } from 'zod'
 
 import { AboutPanel } from '../components/AboutPanel'
+import { DataCreditPanel } from '../components/DataCreditPanel'
 import { IndexPanel } from '../components/IndexPanel'
 import { ArticlePanel } from '../components/article-panel'
 import { ComparisonPanel } from '../components/comparison-panel'
@@ -55,6 +56,7 @@ export type PageDescriptor = ({ kind: 'article' } & z.infer<typeof articleSchema
     | ({ kind: 'random' } & z.infer<typeof randomSchema>)
     | { kind: 'index' }
     | { kind: 'about' }
+    | { kind: 'dataCredit' }
 
 type PageData =
     { kind: 'article', article: Article, universe: string }
@@ -62,6 +64,7 @@ type PageData =
     | { kind: 'statistic', universe: string } & StatisticPanelProps
     | { kind: 'index' }
     | { kind: 'about' }
+    | { kind: 'dataCredit' }
 
 type NavigationState = { state: 'notFound', error: unknown }
     | {
@@ -154,6 +157,9 @@ function urlFromPageDescriptor(pageDescriptor: PageDescriptor): URL {
             pathname = '/about.html'
             searchParams = {}
             break
+        case 'dataCredit':
+            pathname = '/data-credit.html'
+            searchParams = {}
     }
     // eslint-disable-next-line no-restricted-syntax -- Core navigation functions
     const result = new URL(window.location.href)
@@ -282,6 +288,7 @@ async function loadPageDescriptor(descriptor: PageDescriptor, settings: Settings
 
         case 'index':
         case 'about':
+        case 'dataCredit':
             return { pageData: descriptor, newPageDescriptor: descriptor }
     }
 }
@@ -460,5 +467,7 @@ function PageRouter({ pageData }: { pageData: PageData }): ReactNode {
             return <IndexPanel />
         case 'about':
             return <AboutPanel />
+        case 'dataCredit':
+            return <DataCreditPanel />
     }
 }
